1week last

+++ src/main/java/com/ll/gramgram/boundedContext/likeablePerson/controller/LikeablePersonController.java
@@ -64,7 +64,7 @@ public class LikeablePersonController {
 
     @PreAuthorize("isAuthenticated()")
     @GetMapping("/delete/{id}")
-    public String deleteLikePerson(@PathVariable("id")Integer id){
+    public String deleteLikePerson(@PathVariable("id")Long id){
         LikeablePerson likeablePerson = this.likeablePersonService.FindById(id).orElse(null);
         if(likeablePerson == null)
             return rq.historyBack("이미 취소되었습니다.");
